Fixes to remove compiler warnings
Implement fixes for all warnings reported by compiler (GCC 4.3.4 with
-W -Wall flags). These fall into three main categories

- Local scope unused variables, fix by deletion or ignoring return
parameters.

- Unused function arguments, fix by commenting out argument name. This
helps to retain the context of the argument.

- Wrong type comparison (e.g. int -> size_t), fix by using correct type
or cast where behaviour uncertain.

+++ source/bxdatatools/src/library_info.cc
@@ -506,7 +506,7 @@ namespace datatools {
     datatools::library_info::names(libnames);
     out_ << indent << i_tree_dumpable::tag
          << "Registered libraries/components : " << libnames.size() << std::endl;
-    for (int i = 0; i < libnames.size(); i++) {
+    for (size_t i = 0; i < libnames.size(); i++) {
       out_ << indent << i_tree_dumpable::skip_tag;
       if (i < libnames.size() - 1 ) {
         out_ << i_tree_dumpable::tag;
